Create README.md
Yes, it is possible to convert the esp32 encryption/decryption code to Go (also known as Golang). 

There are a few different approaches you can take to do this but I chosen the Manual method.
Manually translate the code: This involves going through the C code and writing the equivalent Go code line by line. This can be time-consuming, but it is a good way to learn Go and understand how the code is working.